// todo/kesq2/src/model.d.ts
export type QueryQuestion = {
	store?: Record<string, true>;
	query?: Record<string, true>;
};
export type QueryAnswer = {
	store: Record<string, StorageValue>;
	query: Record<string, [
		unknown,
		boolean
	]>;
};
export type QueryGenerator<T> = Generator<QueryQuestion, T, QueryAnswer>;
export type QueryFunction<QX, T> = (qx: QX, prev: T | undefined, prevIsValid: boolean) => QueryGenerator<T>;
// ...
export type List = {
	id: string;
	name: string;
	items: string[];
	archived: boolean;
};
export type Item = {
	id: string;
	text: string;
	done: boolean;
	archived: boolean;
};
// ...
declare const TodoQueryContext: {
	get: {
		all_lists: () => QueryGenerator<string[]>;
		item: (item_id: string) => QueryGenerator<Item>;
		list: (list_id: string) => QueryGenerator<List>;
	};
};
export type TodoQX = typeof TodoQueryContext;